requirements.txt
azure-ai-textanalytics==5.3.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
# Novas bibliotecas:
streamlit==1.31.0
pandas==2.2.0
plotly==5.18.0